Apply speedup to cbarT1 and cbarT2
Speedup is achieved by eliminating third derivatives using integration
by parts. The resulting formulas are less elegant-looking, so the
formulas with third derivatives should be kept in the actual text.
Additionally, if we do get tables with third derivatives that we can
interpolate, then there's no longer a speed gain to this IBP trick.