Rename some internal functions
The function renames are
piso -> pressure
pani -> shear

The original rationale was that piso and pani respectively described the
isotropic pressure and pressure anisotropy. This was true for the
unpolarized ensembles (and spin-zero and spin-half composite systems),
but for the tensor-polarized contributions this isn't quite true.
Actually for T1 it is as well, but for T2, the "pressure" function is
strictly anisotropic (it contributes nothing to the trace of the stress
tensor), and the "shear" function does contribute to the isotropic part.

To be sure, it's not quite right to physically call these things
"pressure" and "shear" either, but if it's between an established
misnomer and a novel misnomer, the established nomenclature is better.
It'd however still be a good idea to point out in the text that pT2 and
sT2 are a bit peculiar.