Significantly faster deuteron form factors
- Using integrand formulas directly output by sympy scripts
- Use an mNfm constant instead of manually dividing by hbar
- Don't parallelize the form factor computations

Parallelization is now actually slower because of the overhead created
by posix system calls, so it's been removed.

I don't understand why rewriting the integrands in this way speeds up
the integration so much, but for the cbar-like form factors there's
almost a 10x speedup.

cT1 now appears to be a factor 1/2 smaller. It's likely I mistyped the
formula before. Since I copied and pasted the output this time, there's
no chance of a typo. The formula in the paper draft needs to be checked
however.