Total refactor of density module
The density module is now initialized with the intended bmax and the
number of b points per Cartesian dimension. Additionally, it creates and
saves a cache of any Bessel transform the first time this Bessel
transform is called. This will save time when making density plots by
reducing the number of redundant computations.