Cleanup of EMT-FF code
This removes a lot of the alternate formula codes, leaving only two
choices for formulas for cU, cT1 and cT2. These default to the 'fast'
formula (default) and the 'paper' formula, and are left in place just so
that a reader can verify for themselves that the paper formula and fast
implementation are equivalent.

This commit also removes a lot of code from devplots.py and no longer
automatically imports it. The cbar comparisons are moved into a new
tests.py file. As we finish up the paper, devplots.py will be deprecated
and removed.